requirements.txt
boto3
awscli
boto3
langchain
numpy
langchain_community
python-dotenv
certifi
charset-normalizer
colorama
customtkinter
darkdetect
diffusers
filelock
faiss-cpu
huggingface-hub
idna
importlib-metadata
numpy
packaging
Pillow
pyparsing
PyYAML
regex
requests
tk
tokenizers
torch
torchaudio
torchvision
tqdm
transformers
typing_extensions
zipp
streamlit
boto3
streamlit
pillow
pypdf
google-generativeai
python-dotenv